+++ test/integration/executing_test.exs
@@ -1,6 +1,8 @@
 defmodule Oban.Integration.ExecutingTest do
   use Oban.Case
 
+  import ExUnit.CaptureLog
+
   @moduletag :integration
 
   @oban_opts repo: Repo, queues: [alpha: 3, beta: 3, gamma: 3, delta: 3], poll_interval: 20
@@ -13,33 +15,35 @@ defmodule Oban.Integration.ExecutingTest do
 
   property "individual jobs inserted into running queues are executed" do
     check all jobs <- list_of(job()), max_runs: 20 do
-      jobs = Enum.map(jobs, &Oban.insert!/1)
+      capture_log(fn ->
+        jobs = Enum.map(jobs, &Oban.insert!/1)
 
-      for job <- jobs do
-        %{args: %{ref: ref, action: action}, id: id, max_attempts: max} = job
+        for job <- jobs do
+          %{args: %{ref: ref, action: action}, id: id, max_attempts: max} = job
 
-        assert_receive {_, ^ref}
+          assert_receive {_, ^ref}
 
-        with_backoff(fn ->
-          job = Repo.get(Job, id)
+          with_backoff(fn ->
+            job = Repo.get(Job, id)
 
-          assert job.attempt == 1
-          assert job.attempted_at
-          assert job.completed_at
-          assert job.state == action_to_state(action, max)
+            assert job.attempt == 1
+            assert job.attempted_at
+            assert job.completed_at
+            assert job.state == action_to_state(action, max)
 
-          if job.state != "completed" do
-            assert length(job.errors) > 0
-            assert [%{"attempt" => 1, "at" => _, "error" => _} | _] = job.errors
-          end
-        end)
-      end
+            if job.state != "completed" do
+              assert length(job.errors) > 0
+              assert [%{"attempt" => 1, "at" => _, "error" => _} | _] = job.errors
+            end
+          end)
+        end
+      end)
     end
   end
 
   defp job do
     gen all queue <- member_of(~w(alpha beta gamma delta)),
-            action <- member_of(~w(OK FAIL ERROR EXIT)),
+            action <- member_of(~w(OK FAIL ERROR EXIT TASK_ERROR)),
             ref <- integer(),
             max_attempts <- integer(1..20),
             priority <- integer(0..3),